Support WASG in trace viewer
Summary:
I've developed stockholm syndrome for WASG. So wrote a small diff to support it.

Also changed the color scheme a tiny bit. This isn't intended to be permanent. Just boosted the saturation and value a bit to make debugging more enjoyable.

Test Plan: panned and zoomed aroudn

Reviewers: marcello, leland

Reviewed By: leland

Subscribers: jenkins

Revert Plan: not prod tool

Differential Revision: https://code.uberinternal.com/D1513949