Focus on row after you click on it
Summary:
Focus on a row after you click on it. This diff focuses on the call click detection.

This isn't the full interaction we want. Ideally when you click on a row, it's details should be displayed. And the focusing logic should be more complex (ex: highlight it in the current view if it is already a reasonable scale).

Test Plan:
see video.

{F47825439}

Reviewers: leland, marcello

Reviewed By: leland

Revert Plan: none. not prod.

Differential Revision: https://code.uberinternal.com/D1529265