Trace Visualizer: Selected Crosshairs
Summary:
This is a feature that's sorely missing from Chrome trace. The crosshairs will be particularly useful for navigating through call instances with arrow keys. Search no longer focuses on the selected item and relies on the crosshairs for orientation. Follow up is to improve the situation where a new selection (via search or arrow key navigation) is offscreen.

Meta point:
Moving the screen is disorienting so we should avoid doing so as much as possible.

Test Plan: {F47831773}

Reviewers: attwell

Reviewed By: attwell

Revert Plan: none - not prod app

Differential Revision: https://code.uberinternal.com/D1529313